+++ database/migrations/2021_06_24_142717_create_restaurent_features_table.php
@@ -14,8 +14,8 @@ class CreateRestaurentFeaturesTable extends Migration
     public function up()
     {
         Schema::create('restaurent_features', function (Blueprint $table) {
-            $table->unsignedInteger('restaurent_id');
-            $table->unsignedInteger('feature_id');
+            $table->unsignedBigInteger('restaurent_id');
+            $table->unsignedBigInteger('feature_id');
             $table->timestamps();
 
             $table->foreign('restaurent_id')->references('id')->on('restaurents')->onDelete('cascade');
